pyproject.toml
[project]
name = "geospatial-dvf"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.34",
    "duckdb>=1.4.3",
    "fastapi>=0.128.0",
    "fiona>=1.10.1",
    "geopandas>=1.1.2",
    "google-cloud-storage>=3.8.0",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "polars>=1.37.1",
    "py7zr>=1.1.2",
    "pyarrow>=23.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "uvicorn>=0.40.0",
]
